requirements.txt
slack-sdk>=3.27.0
notion-client>=2.2.1
python-dotenv>=1.0.0
